modify storage libs initialization
